refactor armourhud and inventory viewer
