+++ src/main/java/me/zeroeightsix/kami/module/modules/zeroeightysix/render/ArmourHUD.java
@@ -13,7 +13,7 @@ import net.minecraft.item.ItemStack;
 /**
  * Created by 086 on 24/01/2018.
  */
-@Module.Info(name = "ArmourHUD", category = Module.Category.RENDER)
+@Module.Info(name = "ArmourHUD", category = Module.Category.GUI, showOnArray = Module.ShowOnArray.OFF)
 public class ArmourHUD extends Module {
 
     private static RenderItem itemRender = Minecraft.getMinecraft()
